# tests/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Trajet
  version: 0.4.0
  summary: API del servidor Trajet 0.4.0 — la de la 0.3.0 (compatibilidad) más /api/v1 (iPhone) y /api/admin (panel).
  description: |
    **Contrato congelado de la v2 (FASE 1).** Tres familias de rutas:

    - `/api/*` (0.3.0): se conservan tal cual por compatibilidad, detrás del
      login de Umbrel. La descripción de abajo es la de la 0.3.0; la copia
      exacta de producción está en `docs/openapi-0.3.0.yaml`.
    - `/api/v1/*`: la API de la app del iPhone. Fuera del login de Umbrel
      (`PROXY_AUTH_WHITELIST: "/api/v1/*"`), con token de dispositivo
      obligatorio salvo `/api/v1/ping` y `/api/v1/pair`. Errores con el sobre
      `ErrorV1`. `ETag`/304 y gzip.
    - `/api/admin/*`: el panel, detrás del login de Umbrel, con cabecera
      `X-Trajet-Panel: 1` en lo que modifica.

    Descripción de la API **0.3.0** del servidor Trajet (repo `trajet-server`,
    código de producción 0.3.0), derivada del código Python y no de la
    documentación automática de FastAPI.

    Diferencias con el `/openapi.json` que sirve FastAPI:
    - FastAPI anuncia `info.version: 0.1.0` porque `FastAPI(...)` no recibe
      `version` (app/main.py:43). La versión real de la app es 0.3.0
      (umbrel-app.yml).
    - FastAPI no declara ningún esquema de respuesta (las funciones devuelven
      `dict`) ni de cuerpo (el cuerpo se declara como `payload: dict`). Aquí
...
      diccionario y los de cuerpo de lo que el código lee.
...

...
servers:
...
paths:
...
  /api/routes:
    get:
      tags: [rutas]
      operationId: api_routes_api_routes_get
      summary: Lista de rutas guardadas y la que toca ahora
      responses:
        '200':
          description: Rutas ordenadas por `position, id`.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteList'}
    post:
      tags: [rutas]
      operationId: api_create_route_api_routes_post
      summary: Crear una ruta a mano
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RouteInput'}
      responses:
        '200':
          description: Ruta creada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteSaved'}
        '400':
          description: |
            Falta `name`, `origin_id` o `dest_id` (vacíos tras `strip`), `legs`
            vacío o ausente, o un tramo sin `line_id`/`from_id`
            (app/main.py:94-103).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '422':
          description: El cuerpo no es un objeto JSON.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '500':
          description: |
            Excepción no controlada al guardar: falta `origin_name` o
            `dest_name` (KeyError, app/db.py:212-213), `name` no es texto,
            `days` es null o tiene valores no enteros, `duration_min` o
            `position` no convertibles a entero.
          content:
            text/plain:
              schema: {type: string}
  /api/routes/{route_id}:
    put:
      tags: [rutas]
      operationId: api_update_route_api_routes__route_id__put
      summary: Reemplazar una ruta (cabecera y todos sus tramos)
      parameters:
        - name: route_id
          in: path
          required: true
          schema: {type: integer}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RouteInput'}
      responses:
        '200':
          description: Ruta actualizada. Los tramos se borran y se vuelven a insertar (cambian sus `id`).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteSaved'}
        '400':
          description: Igual que en POST /api/routes.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '404':
          description: '`{"detail": "ruta no encontrada"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '422':
          description: '`route_id` no es entero o el cuerpo no es un objeto JSON.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '500':
          description: Igual que en POST /api/routes.
          content:
            text/plain:
              schema: {type: string}
    delete:
      tags: [rutas]
      operationId: api_delete_route_api_routes__route_id__delete
      summary: Borrar una ruta
      description: Borra la ruta; `ON DELETE CASCADE` borra sus tramos y su historial. Los andenes aprendidos (`platform_obs`) no dependen de la ruta y se quedan.
      parameters:
        - name: route_id
          in: path
          required: true
          schema: {type: integer}
      responses:
        '200':
          description: Borrada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteDeleted'}
        '404':
          description: '`{"detail": "ruta no encontrada"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '422':
          description: '`route_id` no es entero.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
  /api/search/stops:
    get:
      tags: [buscador]
      operationId: api_search_stops_api_search_stops_get
      summary: Buscar zonas de parada (para el editor de rutas)
      description: |
        0 o 1 llamada a Navitia `GET /places?q=…&count=12&type[]=stop_area`
        (cubo de cuota `navitia`). Caché 24 h con clave
        `pl:stop_area:<q en minúsculas y sin espacios en los extremos>`.
      parameters:
        - name: q
          in: query
          required: true
          schema: {type: string, minLength: 2}
      responses:
        '200':
          description: Resultados.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StopSearch'}
        '422':
          description: Falta `q` o tiene menos de 2 caracteres.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '502':
          description: 'PRIM falla y no hay copia en caché: `{"detail": "la API de IDFM no responde: <error>"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
  /api/stops/{stop_id}/lines:
    get:
      tags: [buscador]
      operationId: api_stop_lines_api_stops__stop_id__lines_get
      summary: Líneas que pasan por una zona de parada
      description: |
        0 o 1 llamada a Navitia `GET /stop_areas/{stop_id}/lines?count=60`
        (cubo `navitia`). Caché 24 h, clave `ls:{stop_id}`. Orden: sustitución
        al final, luego modo (metro, RER, tren, TER, tranvía, funicular,
        navette/cable, resto) y código numérico (app/idfm.py:113-147).
      parameters:
        - name: stop_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Líneas.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StopLines'}
        '422':
          description: FastAPI lo declara, pero con un parámetro de ruta de tipo texto no puede darse.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '502':
          description: 'PRIM falla sin caché: `{"detail": "la API de IDFM no responde: <error>"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
  /api/stops/{stop_id}/directions:
    get:
      tags: [buscador]
      operationId: api_stop_directions_api_stops__stop_id__directions_get
      summary: Destinos que circulan ahora para una línea en una parada
      description: |
        0 o 1 llamada a SIRI `stop-monitoring` con
        `MonitoringRef=STIF:StopArea:SP:<código>:` (cubo `stop-monitoring`),
        TTL 25 s, clave `sm:STIF:StopArea:SP:<código>:` (compartida con el
        tablero y el recolector).
      parameters:
        - name: stop_id
          in: path
          required: true
          schema: {type: string}
        - name: line_id
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Destinos ordenados de más a menos frecuente.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Directions'}
        '422':
          description: Falta `line_id`.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '502':
          description: 'PRIM falla sin caché: `{"detail": "la API de IDFM no responde: <error>"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
  /api/board:
    get:
      tags: [tablero]
      operationId: api_board_api_board_get
      summary: Tablero de una ruta (la activa si no se indica)
      description: |
        Llamadas a PRIM: 0 o 1 `stop-monitoring` por ESTACIÓN distinta de la
        ruta (TTL adaptativo 20/30/90/300/600 s, 25 s la primera vez) + 0 o 1
        `general-message` (`LineRef=ALL`, TTL 150 s). Efectos en BD: inserta
        andenes vistos en `platform_obs`, puntuaciones en `platform_score`,
        como mucho una fila en `history` cada 10 min por ruta (si
        `log_history`) y lanza traducciones en segundo plano que acaban en
        `translations`.
      parameters:
        - name: route_id
          in: query
          required: false
          schema: {type: [integer, 'null']}
        - name: log_history
          in: query
          required: false
          schema: {type: boolean, default: true}
      responses:
        '200':
          description: 'Tablero, o `{"empty": true, ...}` si no hay ninguna ruta guardada (aunque se pase `route_id`).'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Board'
                  - $ref: '#/components/schemas/BoardEmpty'
        '404':
          description: '`route_id` no existe (solo si hay alguna ruta): `{"detail": "ruta no encontrada"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '422':
          description: '`route_id` no es entero o `log_history` no es booleano.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '502':
          description: |
            Declarado en app/main.py:185-186, pero `build_board` usa
            `asyncio.gather(..., return_exceptions=True)` y nunca propaga
            `PrimError`: en la práctica los fallos van a `errors` del tablero.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
  /api/alternatives/{route_id}:
    get:
      tags: [alternativas]
      operationId: api_alternatives_api_alternatives__route_id__get
      summary: Itinerarios alternativos evitando las líneas perturbadas
      description: |
        Hasta 3 llamadas a PRIM: `general-message` (TTL 150 s) y, solo si hay
        líneas tocadas o `force=true`, Navitia `journeys` con
        `forbidden_uris[]` (TTL 600 s) y otra `journeys` sin exclusiones para
        la duración de referencia (se guarda 12 h en memoria por ruta).
      parameters:
        - name: route_id
          in: path
          required: true
          schema: {type: integer}
        - name: force
          in: query
          required: false
          schema: {type: boolean, default: false}
      responses:
        '200':
          description: Alternativas, o la respuesta corta si no hace falta ninguna.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Alternatives'
                  - $ref: '#/components/schemas/AlternativesNotNeeded'
        '404':
          description: '`{"detail": "ruta no encontrada"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '422':
          description: '`route_id` no es entero o `force` no es booleano.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '502':
          description: 'La llamada `journeys` con exclusiones falla sin caché: `{"detail": "el calculador no responde: <error>"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
  /api/search/places:
    get:
      tags: [planificador]
      operationId: api_search_places_api_search_places_get
      summary: Buscar paradas, direcciones postales y sitios (planificador)
      description: |
        0 o 1 llamada a Navitia `GET /places?q=…&count=10&type[]=stop_area&type[]=address&type[]=poi`
        (cubo `navitia`). Caché 24 h, clave `pl:stop_area+address+poi:<q normalizada>`.
      parameters:
        - name: q
          in: query
          required: true
          schema: {type: string, minLength: 2}
      responses:
        '200':
          description: Resultados.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlaceSearch'}
        '422':
          description: Falta `q` o tiene menos de 2 caracteres.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '502':
          description: 'PRIM falla sin caché: `{"detail": "la API de IDFM no responde: <error>"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
  /api/plan:
    get:
      tags: [planificador]
      operationId: api_plan_api_plan_get
      summary: Opciones de trayecto puerta a puerta
      description: |
        0 o 1 llamada a Navitia `GET /journeys` con `min_nb_journeys=4`,
        `max_nb_journeys=6`, `data_freshness=realtime` y, si hay `when`,
        `datetime=<YYYYMMDDTHHMMSS>` y `datetime_represents=<mode>`
        (cubo `navitia`). Caché 600 s.
      parameters:
        - name: from
          in: query
          required: true
          description: Id Navitia de origen (`stop_area:IDFM:…`, `lon;lat` de una dirección, poi…).
          schema: {type: string, minLength: 3}
        - name: to
          in: query
          required: true
          description: Id Navitia de destino.
          schema: {type: string, minLength: 3}
        - name: when
          in: query
          required: false
          description: '`HH:MM` en hora de París. Si ya ha pasado hoy, se planifica para mañana. Si no se puede partir en dos enteros se ignora (ahora).'
          schema: {type: [string, 'null']}
        - name: mode
          in: query
          required: false
          description: '`departure` o `arrival` (validado a mano, 400 si no).'
          schema: {type: string, default: departure}
      responses:
        '200':
          description: Opciones ordenadas y sin duplicar por combinación de líneas.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Plan'}
        '400':
          description: '`{"detail": "mode tiene que ser ''departure'' o ''arrival''"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '404':
          description: 'Ninguna opción con transporte público: `{"detail": "no hay ningún trayecto en transporte público entre esos dos puntos a esa hora"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '422':
          description: Falta `from` o `to`, o tiene menos de 3 caracteres.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '500':
          description: '`when` con hora o minuto fuera de rango (p. ej. `25:00`): `datetime.replace` lanza ValueError fuera del try (app/planner.py:216).'
          content:
            text/plain:
              schema: {type: string}
        '502':
          description: 'PRIM falla sin caché: `{"detail": "la API de IDFM no responde: <error>"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
  /api/routes/from-plan:
    post:
      tags: [planificador, rutas]
      operationId: api_route_from_plan_api_routes_from_plan_post
      summary: Guardar como ruta vigilada una opción del planificador
      description: |
        Por cada tramo con `from_id`, `line_id` y `direction` no vacíos, 0 o 1
        llamada a SIRI `stop-monitoring` de su estación (TTL 25 s; dos tramos
        de la misma estación comparten llamada) para traducir la dirección de
        Navitia al destino que usa el tiempo real. NO pasa por
        `_validate_route`: se guarda lo que llegue.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RouteFromPlanInput'}
      responses:
        '200':
          description: Ruta creada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteFromPlanSaved'}
        '400':
          description: '`option.legs` vacío o ausente: `{"detail": "falta el itinerario elegido"}`'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorDetail'}
        '422':
          description: El cuerpo no es un objeto JSON.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
        '500':
          description: '`option` o `meta` no son objetos, un tramo no es objeto, `days` con valores no enteros, `minutes` no convertible a entero…'
          content:
            text/plain:
              schema: {type: string}
  /api/platform-model:
    get:
      tags: [estadisticas]
      operationId: api_platform_model_api_platform_model_get
      summary: Qué sabe la previsión de andén y si acierta
      description: No gasta cuota. `route_id` omitido = ruta activa. Un `route_id` inexistente (o 0) no da error; simplemente no se incluyen `route` ni `coverage`.
      parameters:
        - name: route_id
          in: query
          required: false
          schema: {type: [integer, 'null']}
      responses:
        '200':
          description: Modelo.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlatformModel'}
        '422':
          description: '`route_id` no es entero.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
  /api/stats:
    get:
      tags: [estadisticas]
      operationId: api_stats_api_stats_get
      summary: Historial de incidencias
      description: No gasta cuota. Tres consultas sobre `history` filtradas por `day >= date('now', '-<days> days')`.
      parameters:
        - name: days
          in: query
          required: false
          schema: {type: integer, default: 90}
      responses:
        '200':
          description: Resumen.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Stats'}
        '422':
          description: '`days` no es entero.'
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HTTPValidationError'}
...
  /api/v1/board:
    get:
      tags: [v1-tablero]
      operationId: v1_board
      security: [{deviceToken: []}]
      summary: Tablero de una ruta (la que toca si no se indica)
      description: |
        Igual que `/api/board` más el estado del servidor (`server`) y los
        campos `from_id`, `to_id` y `platform_expected` por tramo.

        **Nunca se manda un tablero hueco.** Si no se ha podido traer NADA
        (ninguna estación ni avisos, ni en caché) se responde con error
        (`prim_key_missing` 503, `prim_quota_exhausted` 503,
        `prim_key_invalid` 503 o `prim_unreachable` 502) para que la app se
        quede con su último tablero bueno y enseñe el estado diseñado. Si
        solo falla una parte, 200 con esa parte en `errors`.

        Acepta `If-None-Match` y responde 304 si el cuerpo no ha cambiado.
        `server.refresh_hint_s` dice cada cuánto conviene volver a pedirlo
        (30 s normalmente; más si la cuota va justa).
      parameters:
        - name: route_id
          in: query
          required: false
          schema: {type: [integer, 'null']}
        - name: log_history
          in: query
          required: false
          description: La app manda `false` en los refrescos automáticos (R40).
          schema: {type: boolean, default: true}
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Tablero, o vacío si no hay rutas.
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/BoardV1'
                  - $ref: '#/components/schemas/BoardEmptyV1'
        '304': {$ref: '#/components/responses/NotModified'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '502': {$ref: '#/components/responses/Upstream'}
        '503': {$ref: '#/components/responses/PrimUnavailable'}
  /api/v1/routes:
    get:
      tags: [v1-rutas]
      operationId: v1_routes_list
      security: [{deviceToken: []}]
      summary: Rutas guardadas y la que toca ahora
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Rutas.
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteList'}
        '304': {$ref: '#/components/responses/NotModified'}
        '401': {$ref: '#/components/responses/Unauthorized'}
    post:
      tags: [v1-rutas]
      operationId: v1_routes_create
      security: [{deviceToken: []}]
      summary: Crear una ruta a mano
      description: Validación completa (400 con mensaje claro, nunca 500).
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RouteInput'}
      responses:
        '201':
          description: Creada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteSaved'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /api/v1/routes/{route_id}:
    parameters:
      - name: route_id
        in: path
        required: true
        schema: {type: integer}
    get:
      tags: [v1-rutas]
      operationId: v1_routes_get
      security: [{deviceToken: []}]
      summary: Una ruta
      responses:
        '200':
          description: Ruta.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Route'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      tags: [v1-rutas]
      operationId: v1_routes_update
      security: [{deviceToken: []}]
      summary: Reemplazar una ruta (cabecera y tramos)
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RouteInput'}
      responses:
        '200':
          description: Actualizada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteSaved'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [v1-rutas]
      operationId: v1_routes_delete
      security: [{deviceToken: []}]
      summary: Borrar una ruta (y su historial; los andenes aprendidos se quedan)
      responses:
        '200':
          description: Borrada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteDeleted'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/routes/from-plan:
    post:
      tags: [v1-rutas]
      operationId: v1_routes_from_plan
      security: [{deviceToken: []}]
      summary: Guardar como ruta una opción del planificador
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RouteFromPlanInput'}
      responses:
        '201':
          description: Creada.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RouteFromPlanSaved'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
  /api/v1/search/stops:
    get:
      tags: [v1-buscador]
      operationId: v1_search_stops
      security: [{deviceToken: []}]
      summary: Buscar zonas de parada (editor manual)
      parameters:
        - name: q
          in: query
          required: true
          schema: {type: string, minLength: 2, maxLength: 80}
      responses:
        '200':
          description: Resultados.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StopSearch'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '502': {$ref: '#/components/responses/Upstream'}
        '503': {$ref: '#/components/responses/PrimUnavailable'}
  /api/v1/search/places:
    get:
      tags: [v1-buscador]
      operationId: v1_search_places
      security: [{deviceToken: []}]
      summary: Buscar paradas, direcciones y sitios (planificador)
      parameters:
        - name: q
          in: query
          required: true
          schema: {type: string, minLength: 2, maxLength: 80}
      responses:
        '200':
          description: Resultados.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlaceSearch'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '502': {$ref: '#/components/responses/Upstream'}
        '503': {$ref: '#/components/responses/PrimUnavailable'}
  /api/v1/stops/{stop_id}/lines:
    get:
      tags: [v1-buscador]
      operationId: v1_stop_lines
      security: [{deviceToken: []}]
      summary: Líneas que pasan por una zona de parada (ordenadas por modo)
      parameters:
        - name: stop_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Líneas.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StopLines'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '502': {$ref: '#/components/responses/Upstream'}
        '503': {$ref: '#/components/responses/PrimUnavailable'}
  /api/v1/stops/{stop_id}/directions:
    get:
      tags: [v1-buscador]
      operationId: v1_stop_directions
      security: [{deviceToken: []}]
      summary: Destinos que circulan ahora para una línea en una parada
      parameters:
        - name: stop_id
          in: path
          required: true
          schema: {type: string}
        - name: line_id
          in: query
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Destinos.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Directions'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '502': {$ref: '#/components/responses/Upstream'}
        '503': {$ref: '#/components/responses/PrimUnavailable'}
...
  /api/v1/stats:
    get:
      tags: [v1-estadisticas]
      operationId: v1_stats
      security: [{deviceToken: []}]
      summary: Historial de incidencias
      parameters:
        - name: days
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 3650, default: 90}
      responses:
        '200':
          description: Resumen.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StatsV1'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
...
components:
  schemas:
...
    RouteDeleted:
      type: object
      additionalProperties: false
      required: [deleted]
      properties:
        deleted: {type: integer}
...
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      description: 'Token de dispositivo obtenido en /api/v1/pair.'